# public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "version": "1.0.0",
    "title": "Book-a-Place API",
    "description": "A simple API for an Airbnb-like application.\n"
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ],
  "paths": {
    "/lodgings": {
      "get": {
        "summary": "Fetch a list of lodgings",
        "description": "Returns a paginated list of lodgings.\n",
        "operationId": "getLodgings",
        "tags": [
          "lodgings"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Specifies a specific page of lodgings to request.\n",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pageNumber": {
                      "type": "integer",
                      "description": "The current page of lodgings.",
                      "example": 1
                    },
                    "totalPages": {
                      "type": "integer",
                      "description": "Total pages of lodgings.",
                      "example": 127
                    },
                    "lodgings": {
                      "type": "array",
                      "description": "The returned lodgings.",
                      "items": {
                        "$ref": "#/components/schemas/Lodging"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a new lodging.",
        "description": "Creates a new lodging with specified data.\n",
        "operationId": "createLodging",
        "tags": [
          "lodgings"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Lodging"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "integer",
                      "format": "int64",
                      "description": "ID of the created lodging.",
                      "example": 12345
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Incorrectly-formatted request body.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/lodgings/{lodgingID}": {
      "parameters": [
        {
          "name": "lodgingID",
          "in": "path",
          "description": "ID of the lodging.",
          "schema": {
            "type": "integer"
          },
          "required": true
        }
      ]
    }
  },
  "components": {
    "schemas": {
      "Lodging": {
        "description": "An object representing a single lodging.\n",
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Name of the lodging.",
            "example": "My Cool Condo"
          },
          "description": {
            "type": "string",
            "description": "Textual description of the lodging.",
            "example": "Nice place to stay, down by the river."
          },
          "price": {
            "type": "number",
            "description": "Price per night of the lodging.",
            "example": 128,
            "minimum": 0.01
          }
        },
        "required": [
          "name",
          "price"
        ]
      },
      "Error": {
        "description": "Object representing an API error.\n",
        "type": "object",
        "properties": {
          "err": {
            "type": "string",
            "description": "Error message."
          }
        }
      }
    }
  }
}